requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==21.6b0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==3.2.7
django-allauth==0.54.0
django-anymail==10.0
django-autoslug==1.9.8
django-cors-headers==3.14.0
django-countries==7.2.1
django-environ==0.7.0
django-filter==21.1
django-otp==1.2.2
django-phonenumber-field==5.2.0
django-smtp-ssl==1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.21.2
drf-spectacular-sidecar==2023.7.1
drf-yasg==1.21.6
-e git+https://github.com/muepsilon/elasticemail-django@1dad468041bf0e4e9bf9ac8260fcbf95c8c58a22#egg=elasticemail_django
flake8==3.9.2
frozenlist==1.3.3
gunicorn==20.0.3
idna==3.4
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mccabe==0.6.1
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.27.5
packaging==23.1
pathspec==0.11.1
phonenumbers==8.12.33
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
pyotp==2.8.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
sib-api-v3-sdk==7.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.9.2
